// src/data/gen/units.ts
/**
 * @license
 * Copyright 2026 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

// @generated by scripts/codegen/generate_units_code.ts. Do not edit.
// Machine-generated.

import type { CSSNumericType } from '../../typed-om.ts';

export const UNITS = [
  'cap',
  'ch',
  'cm',
  'cqb',
  'cqh',
  'cqi',
  'cqmax',
  'cqmin',
  'cqw',
  'deg',
  'dpcm',
  'dpi',
  'dppx',
  'dvb',
  'dvh',
  'dvi',
  'dvmax',
  'dvmin',
  'dvw',
  'em',
  'ex',
  'fr',
  'grad',
  'hz',
  'ic',
  'in',
  'khz',
  'lh',
  'lvb',
  'lvh',
  'lvi',
  'lvmax',
  'lvmin',
  'lvw',
  'mm',
  'ms',
  'number',
  'pc',
  'percent',
  'pt',
  'px',
  'q',
  'rad',
  'rcap',
  'rch',
  'rem',
  'rex',
  'ric',
  'rlh',
  's',
  'svb',
  'svh',
  'svi',
  'svmax',
  'svmin',
  'svw',
  'turn',
  'vb',
  'vh',
  'vi',
  'vmax',
  'vmin',
  'vw',
  'x',
] as const;

export type CSSUnit = typeof UNITS[number];

export const unitToBase: Record<string, keyof CSSNumericType | 'number'> = {
  'cap': 'length',
  'ch': 'length',
  'cm': 'length',
  'deg': 'angle',
  'dpcm': 'resolution',
  'dpi': 'resolution',
  'dppx': 'resolution',
  'em': 'length',
  'ex': 'length',
  'fr': 'flex',
  'grad': 'angle',
  'hz': 'frequency',
  'ic': 'length',
  'in': 'length',
  'khz': 'frequency',
  'mm': 'length',
  'ms': 'time',
  'pc': 'length',
  'pt': 'length',
  'px': 'length',
  'q': 'length',
  'rad': 'angle',
  'rem': 'length',
  's': 'time',
  'turn': 'angle',
  'vh': 'length',
  'vmax': 'length',
  'vmin': 'length',
  'vw': 'length',
  'x': 'resolution',
  'svw': 'length',
  'lvw': 'length',
  'dvw': 'length',
  'svh': 'length',
  'lvh': 'length',
  'dvh': 'length',
  'svmin': 'length',
  'lvmin': 'length',
  'dvmin': 'length',
  'svmax': 'length',
  'lvmax': 'length',
  'dvmax': 'length',
  'lh': 'length',
  'rlh': 'length',
  'vi': 'length',
  'vb': 'length',
  'rex': 'length',
  'rcap': 'length',
  'rch': 'length',
  'ric': 'length',
  'svi': 'length',
  'svb': 'length',
  'lvi': 'length',
  'lvb': 'length',
  'dvi': 'length',
  'dvb': 'length',
  'cqw': 'length',
  'cqh': 'length',
  'cqi': 'length',
  'cqb': 'length',
  'cqmin': 'length',
  'cqmax': 'length',
  'percent': 'percent',
  'number': 'number',
};

export const unitToPixels: Record<string, number> = {
  'cm': 37.79527559055118,
  'in': 96,
  'mm': 3.7795275590551185,
  'pc': 16,
  'pt': 1.3333333333333333,
  'px': 1,
  'q': 0.9448818897637794,
};

export const unitToRadians: Record<string, number> = {
  'deg': 0.017453292519943295,
  'grad': 0.015707963267948967,
  'rad': 1,
  'turn': 6.283185307179586,
};

export const unitToSeconds: Record<string, number> = {
  'ms': 0.001,
  's': 1,
};